requirements.txt
openai==0.27.9
tqdm
requests
aiohttp
python-dotenv